#88 Refactoring (Add subfolders for keepers and teams components and model)
