+++ src/app/app.module.ts
@@ -6,15 +6,15 @@ import { HttpModule } from '@angular/http';
 import { AppComponent } from './app.component';
 import { AllUsersTableComponent } from './components/gamification/all-users-table/all-users-table.component';
 import { UserDetailsTableComponent } from './components/gamification/user-details-table/user-details-table.component';
-import { ActiveKeepersComponent } from './components/gamification/active-keepers/active-keepers.component';
-import { TeamsComponent } from './components/gamification/teams/teams.component';
+import { ActiveKeepersComponent } from './components/keepers/active-keepers.component';
+import { TeamsComponent } from './components/teams/teams.component';
 import { AllChannelsTableComponent } from './components/slack-archive/all-channels-table/all-channels-table.component';
 import { ChannelMessagesTableComponent } from './components/slack-archive/channel-messages-table/channel-messages-table.component';
-import { HeaderComponent } from './components/header/header.component';
-import { GoTopComponent } from './components/go-to-top/go-to-top.component';
+import { HeaderComponent } from './components/shared/header/header.component';
+import { GoTopComponent } from './components/shared/go-to-top/go-to-top.component';
 
 import { GamificationService } from './service/gamification.service';
-import { UserService } from './service/user.service';
+import { UserService } from './service/users.service';
 import { ScrollDirective } from './directives/scroll.directive';
 import { KeepersService } from './service/keepers.service';
 import { TeamsService } from './service/teams.service';
