+++ src/app/components/shared/header/header.component.html
@@ -2,7 +2,7 @@
   <div class="container-fluid">
     <div class="navbar-header">
       <a href="https://juja.com.ua/" class="navbar-brand">
-        <img src="../../../assets/images/brand_logo.png">
+        <img src="../../../../assets/images/brand_logo.png">
       </a>
       <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#menu">
         <span class="icon-bar"></span>
