+++ src/app/model/keepers/keepersDetails.ts
